✨ Réorganisation des paramètres
