+++ src/theme/theme.css
@@ -326,10 +326,11 @@ ion-list.list-inset ion-list-header, ion-list.list-inset ion-item {
 ion-list.list-inset.ios ion-list-header ion-label {
     margin-top: 0px;
     padding-top: 5px;
+    padding-bottom: 5px;
 }
 
 ion-list.list-inset.md ion-list-header {
-    padding-top: 0px;
+    padding-top: 8px;
 }
 
 /* action sheet */
